requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
ultralytics==8.0.196
gspread==5.12.0
oauth2client==4.1.3
pyserial==3.5
gevent==23.9.1
requests

